requirements.txt
langchain
langchain-community
langchain-chroma
langchain_groq
chromadb
streamlit
pandas
numpy
python-dotenv
sentence-transformers
langchain_huggingface